LayoutTests/html5lib/runner-expected.txt

ALERT: Invalid test: #errors
#document-fragment
html
#document
| <head>
| <body>

CONSOLE MESSAGE: line 2: PASS
CONSOLE MESSAGE: line 2: FOO<span>BAR</span>BAZ
ALERT: Invalid test: #errors
#document-fragment
html
#document
| <head>
| <body>

CONSOLE MESSAGE: line 3: PASS
CONSOLE MESSAGE: line 3: FOO<span>BAR</span>BAZ
resources/adoption01.dat:
14

Test 14 of 17 in resources/adoption01.dat failed. Input:
<div><a><b><div><div><div><div><div><div><div><div><div><div></a>
Got:
| <html>
|   <head>
|   <body>
|     <div>
|       <a>
|         <b>
|       <b>
|         <div>
|           <a>
|           <div>
|             <a>
|             <div>
|               <a>
|               <div>
|                 <a>
|                 <div>
|                   <a>
|                   <div>
|                     <a>
|                     <div>
|                       <a>
|                       <div>
|                         <a>
|                           <div>
|                             <div>
Expected:
| <html>
|   <head>
|   <body>
|     <div>
|       <a>
|         <b>
|       <b>
|         <div>
|           <a>
|           <div>
|             <a>
|             <div>
|               <a>
|               <div>
|                 <a>
|                 <div>
|                   <a>
|                   <div>
|                     <a>
|                     <div>
|                       <a>
|                       <div>
|                         <a>
|                         <div>
|                           <div>
resources/adoption02.dat: PASS

resources/comments01.dat: PASS

resources/doctype01.dat: PASS

resources/domjs-unsafe.dat: PASS

resources/entities01.dat: PASS

resources/entities02.dat: PASS

resources/html5test-com.dat: PASS

resources/inbody01.dat: PASS

resources/isindex.dat: PASS

resources/pending-spec-changes-plain-text-unsafe.dat: PASS

resources/pending-spec-changes.dat: PASS

resources/plain-text-unsafe.dat: PASS

resources/scriptdata01.dat: PASS

resources/scripted/adoption01.dat: PASS

resources/scripted/ark.dat: PASS

resources/scripted/webkit01.dat: PASS

resources/tables01.dat: PASS

resources/tests_innerHTML_1.dat: PASS

resources/tricky01.dat: PASS

resources/webkit01.dat: PASS

resources/webkit02.dat: PASS

Switching to data: url parser (threaded if available) and re-running...

resources/adoption01.dat:
14

Test 14 of 17 in resources/adoption01.dat failed. Input:
<div><a><b><div><div><div><div><div><div><div><div><div><div></a>
Got:
| <html>
|   <head>
|   <body>
|     <div>
|       <a>
|         <b>
|       <b>
|         <div>
|           <a>
|           <div>
|             <a>
|             <div>
|               <a>
|               <div>
|                 <a>
|                 <div>
|                   <a>
|                   <div>
|                     <a>
|                     <div>
|                       <a>
|                       <div>
|                         <a>
|                           <div>
|                             <div>
Expected:
| <html>
|   <head>
|   <body>
|     <div>
|       <a>
|         <b>
|       <b>
|         <div>
|           <a>
|           <div>
|             <a>
|             <div>
|               <a>
|               <div>
|                 <a>
|                 <div>
|                   <a>
|                   <div>
|                     <a>
|                     <div>
|                       <a>
|                       <div>
|                         <a>
|                         <div>
|                           <div>
resources/adoption02.dat: PASS

resources/comments01.dat: PASS

resources/doctype01.dat: PASS

resources/domjs-unsafe.dat: PASS

resources/entities01.dat: PASS

resources/entities02.dat: PASS

resources/html5test-com.dat: PASS

resources/inbody01.dat: PASS

resources/isindex.dat: PASS

resources/pending-spec-changes-plain-text-unsafe.dat: PASS

resources/pending-spec-changes.dat: PASS

resources/plain-text-unsafe.dat: PASS

resources/scriptdata01.dat: PASS

resources/scripted/adoption01.dat: PASS

resources/scripted/ark.dat: PASS

resources/scripted/webkit01.dat: PASS

resources/tables01.dat: PASS

resources/tests_innerHTML_1.dat: PASS

resources/tricky01.dat: PASS

resources/webkit01.dat: PASS

resources/webkit02.dat: PASS
